+++ doubleml/utils/tests/test_blp.py
@@ -4,7 +4,6 @@ import pytest
 import copy
 
 import doubleml as dml
-
 from ._utils_blp_manual import fit_blp, blp_confint
