+++ doubleml/utils/tests/test_policytree.py
@@ -5,7 +5,7 @@ import copy
 
 import doubleml as dml
 
-from ._utils_pt_manual import fit_policytree
+from ...tests._utils_pt_manual import fit_policytree
 from sklearn.tree import DecisionTreeClassifier
 from sklearn.exceptions import NotFittedError
